functions: Implement dynamic collection names and database URL
Original app (from tutorial) had static collection name, "items". Made
this dynamic by using getting name from request URL.

Trouble deciding what to call a single entry. Firebase uses "document"
but not sure if this is a generic enough term. It's "doc" for now.

Another undecided is use of `body` const in index.js, as this is might be
better done in a separate config file. It is the only part of this app
which document info specific to a particular Firebase app.

Database URL simply changed from static to dynamic. Uses permissions.json
file, as this contains app ID, which is part of URL.

Had to include meaningless return statement to satisfy lint. Replace this
later.